fix: flee/shy cleanup

+++ common/src/main/java/com/t2pellet/strawgolem/entity/goals/golem/GolemFleeEntityGoal.java
@@ -6,34 +6,44 @@ import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.ai.goal.AvoidEntityGoal;
 import net.minecraft.world.entity.animal.Animal;
 import net.minecraft.world.entity.monster.Monster;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.entity.raid.Raider;
 
 public class GolemFleeEntityGoal<T extends LivingEntity> extends AvoidEntityGoal<T> {
 
     private final StrawGolem golem;
+    private final boolean panic;
 
-    public GolemFleeEntityGoal(StrawGolem mob, Class<T> clazz, float distance, double walkSpeed, double sprintSpeed) {
+    public GolemFleeEntityGoal(StrawGolem mob, Class<T> clazz, float distance, double walkSpeed, double sprintSpeed, boolean panic) {
         super(mob, clazz, distance, walkSpeed, sprintSpeed);
         this.golem = mob;
+        this.panic = panic;
     }
 
     @Override
     public boolean canUse() {
-        if (Monster.class.isAssignableFrom(avoidClass)) {
-            return StrawgolemConfig.Behaviour.golemsRunFromMonsters.get() && super.canUse();
-        } else if (Raider.class.isAssignableFrom(avoidClass)) {
-            return StrawgolemConfig.Behaviour.golemsRunFromRaiders.get() && super.canUse();
-        } else if (Animal.class.isAssignableFrom(avoidClass)) {
-            return StrawgolemConfig.Behaviour.golemsRunFromLivestock.get() && super.canUse();
+        if (!super.canUse() || toAvoid == null) return false;
+
+        Class<?> targetClass = toAvoid.getClass();
+        if (Monster.class.isAssignableFrom(targetClass)) {
+            return StrawgolemConfig.Behaviour.golemsRunFromMonsters.get();
+        } else if (Raider.class.isAssignableFrom(targetClass)) {
+            return StrawgolemConfig.Behaviour.golemsRunFromRaiders.get();
+        } else if (Animal.class.isAssignableFrom(targetClass)) {
+            return StrawgolemConfig.Behaviour.golemsRunFromLivestock.get();
+        } else if (Player.class.isAssignableFrom(targetClass)) {
+            return StrawgolemConfig.Behaviour.golemsRunFromPlayers.get() && !toAvoid.isHolding(StrawGolem.REPAIR_ITEM);
         }
-        return super.canUse();
+        return true;
     }
 
     @Override
     public void start() {
         super.start();
-        golem.getHeldItem().drop();
-        golem.setScared(true);
+        if (panic) {
+            golem.getHeldItem().drop();
+            golem.setScared(true);
+        }
     }
 
     @Override
